refactor to command package
