+++ fabric/src/main/java/com/discordsrv/fabric/command/game/sender/FabricCommandSender.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.discordsrv.fabric.game.sender;
+package com.discordsrv.fabric.command.game.sender;
 
 import com.discordsrv.fabric.FabricDiscordSRV;
 import com.discordsrv.common.command.game.abstraction.sender.ICommandSender;
