+++ fabric/src/main/java/com/discordsrv/fabric/player/FabricPlayer.java
@@ -19,7 +19,7 @@
 package com.discordsrv.fabric.player;
 
 import com.discordsrv.fabric.FabricDiscordSRV;
-import com.discordsrv.fabric.game.sender.FabricCommandSender;
+import com.discordsrv.fabric.command.game.sender.FabricCommandSender;
 import com.discordsrv.common.DiscordSRV;
 import com.discordsrv.common.abstraction.player.IPlayer;
 import com.discordsrv.common.abstraction.player.provider.model.SkinInfo;
